[2.Use Dapr to discover & invoke uservices]
2 - In the sales_tax_rate service, we will read zip code and tax rate data from the CSV file and import it into a key-value store through the Dapr sidecar (i.e., the client). As you can see, we need to structure the data into key/value JSON objects before we can save them.

3 - Then, in the find_rate handler, look up the zip code from the key-value store through the Dapr sidecar.

Deliverable

The deliverable for this milestone is a complete application refactored to use the Dapr state store to manage the zip code and tax rate data table.

Upload a link to your deliverable in the Submit Your Work section and click submit. After submitting, the author’s solution and peer solutions will appear on the page for you to examine.